+++ test/_init.js
@@ -1,3 +1,4 @@
+require("babel-polyfill");
 import log from "../src/log";
 
 // Supress log output.